+++ src/main/java/com/allsimon/intellij/MyMessageBundle.java
@@ -5,7 +5,7 @@ import java.util.function.Supplier;
 import org.jetbrains.annotations.Nls;
 import org.jetbrains.annotations.PropertyKey;
 
-final class MyMessageBundle {
+public final class MyMessageBundle {
     private static final String BUNDLE = "messages.MyMessageBundle";
 
     private static final DynamicBundle INSTANCE = new DynamicBundle(MyMessageBundle.class, BUNDLE);
@@ -13,11 +13,11 @@ final class MyMessageBundle {
     private MyMessageBundle() {
     }
 
-    static @Nls String message(@PropertyKey(resourceBundle = BUNDLE) String key, Object... params) {
+    public static @Nls String message(@PropertyKey(resourceBundle = BUNDLE) String key, Object... params) {
         return INSTANCE.getMessage(key, params);
     }
 
-    static Supplier<@Nls String> lazyMessage(@PropertyKey(resourceBundle = BUNDLE) String key, Object... params) {
+    public static Supplier<@Nls String> lazyMessage(@PropertyKey(resourceBundle = BUNDLE) String key, Object... params) {
         return INSTANCE.getLazyMessage(key, params);
     }
 }